Refactor in terms of a ShaderWrapper object
