+++ manimlib/mobject/mobject.py
@@ -22,7 +22,7 @@ from manimlib.utils.simple_functions import get_parameters
 from manimlib.utils.space_ops import angle_of_vector
 from manimlib.utils.space_ops import get_norm
 from manimlib.utils.space_ops import rotation_matrix_transpose
-from manimlib.utils.shaders import ShaderWrapper
+from manimlib.shader_wrapper import ShaderWrapper
 
 
 # TODO: Explain array_attrs
